+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/GovSpec.hs
@@ -84,15 +84,14 @@ unknownCostModelsSpec =
     it "Are accepted" $ do
       costModels <- getsPParams ppCostModelsL
       newCostModels <- arbitrary
-      (hotCommitteeC1 :| [hotCommitteeC2]) <- registerInitialCommittee
+      hotCommitteeCs <- registerInitialCommittee
       (drepC, _, _) <- setupSingleDRep 1_000_000
       gai <-
         submitParameterChange SNothing $
           emptyPParamsUpdate
             & ppuCostModelsL .~ SJust newCostModels
       submitYesVote_ (DRepVoter drepC) gai
-      submitYesVote_ (CommitteeVoter hotCommitteeC1) gai
-      submitYesVote_ (CommitteeVoter hotCommitteeC2) gai
+      submitYesVoteCCs_ hotCommitteeCs gai
       passNEpochs 2
       getLastEnactedParameterChange `shouldReturn` SJust (GovPurposeId gai)
       getsPParams ppCostModelsL `shouldReturn` updateCostModels costModels newCostModels
@@ -311,7 +310,7 @@ proposalsWithVotingSpec =
                          , SJust <$> b
                          ]
       it "Subtrees are pruned when competing proposals are enacted over multiple rounds" $ do
-        (committeeMember1 :| [committeeMember2]) <- registerInitialCommittee
+        committeeMembers' <- registerInitialCommittee
         (drepC, _, _) <- setupSingleDRep 1_000_000
         a@[ c
             , Node
@@ -339,14 +338,11 @@ proposalsWithVotingSpec =
             , Node () []
             ]
         submitYesVote_ (DRepVoter drepC) p2
-        submitYesVote_ (CommitteeVoter committeeMember1) p2
-        submitYesVote_ (CommitteeVoter committeeMember2) p2
+        submitYesVoteCCs_ committeeMembers' p2
         submitYesVote_ (DRepVoter drepC) p21
-        submitYesVote_ (CommitteeVoter committeeMember1) p21
-        submitYesVote_ (CommitteeVoter committeeMember2) p21
+        submitYesVoteCCs_ committeeMembers' p21
         submitYesVote_ (DRepVoter drepC) p3
-        submitYesVote_ (CommitteeVoter committeeMember1) p3 -- Two competing proposals break the tie based on proposal order
-        submitYesVote_ (CommitteeVoter committeeMember2) p3
+        submitYesVoteCCs_ committeeMembers' p3 -- Two competing proposals break the tie based on proposal order
         fmap (!! 3) getProposalsForest
           `shouldReturn` Node SNothing (fmap SJust <$> a)
         passEpoch
@@ -385,8 +381,7 @@ proposalsWithVotingSpec =
         p2131 <- submitConstitutionGovAction $ SJust p213
         p2141 <- submitConstitutionGovAction $ SJust p214
         submitYesVote_ (DRepVoter drepC) p212
-        submitYesVote_ (CommitteeVoter committeeMember1) p212
-        submitYesVote_ (CommitteeVoter committeeMember2) p212
+        submitYesVoteCCs_ committeeMembers' p212
         fmap (!! 3) getProposalsForest
           `shouldReturn` Node
             (SJust p2)
@@ -406,7 +401,7 @@ proposalsWithVotingSpec =
         proposalsSize props `shouldBe` 0
       it "Votes from subsequent epochs are considered for ratification" $ do
         modifyPParams $ ppGovActionLifetimeL .~ EpochInterval 4
-        (committeeMember1 :| [committeeMember2]) <- registerInitialCommittee
+        committeeMembers' <- registerInitialCommittee
         (dRep, _, _) <- setupSingleDRep 1_000_000
         [Node p1 []] <-
           submitConstitutionForest
@@ -416,13 +411,12 @@ proposalsWithVotingSpec =
           `shouldReturn` Node SNothing [Node (SJust p1) []]
         passNEpochs 2
         submitYesVote_ (DRepVoter dRep) p1
-        submitYesVote_ (CommitteeVoter committeeMember1) p1
-        submitYesVote_ (CommitteeVoter committeeMember2) p1
+        submitYesVoteCCs_ committeeMembers' p1
         passNEpochs 2
         fmap (!! 3) getProposalsForest
           `shouldReturn` Node (SJust p1) []
       it "Subtrees are pruned for both enactment and expiry over multiple rounds" $ do
-        (committeeMember1 :| [committeeMember2]) <- registerInitialCommittee
+        committeeMembers' <- registerInitialCommittee
         (dRep, _, _) <- setupSingleDRep 1_000_000
         modifyPParams $ ppGovActionLifetimeL .~ EpochInterval 4
         [ a@( Node
@@ -461,14 +455,11 @@ proposalsWithVotingSpec =
             ]
         passNEpochs 2
         submitYesVote_ (DRepVoter dRep) p1
-        submitYesVote_ (CommitteeVoter committeeMember1) p1
-        submitYesVote_ (CommitteeVoter committeeMember2) p1
+        submitYesVoteCCs_ committeeMembers' p1
         submitYesVote_ (DRepVoter dRep) p11
-        submitYesVote_ (CommitteeVoter committeeMember1) p11
-        submitYesVote_ (CommitteeVoter committeeMember2) p11
+        submitYesVoteCCs_ committeeMembers' p11
         submitYesVote_ (DRepVoter dRep) p3
-        submitYesVote_ (CommitteeVoter committeeMember1) p3 -- Two competing proposals break the tie based on proposal order
-        submitYesVote_ (CommitteeVoter committeeMember2) p3
+        submitYesVoteCCs_ committeeMembers' p3 -- Two competing proposals break the tie based on proposal order
         passNEpochs 2
         fmap (!! 3) getProposalsForest
           `shouldReturn` SJust
@@ -505,8 +496,7 @@ proposalsWithVotingSpec =
         fmap (!! 3) getProposalsForest
           `shouldReturn` Node (SJust p11) (fmap SJust <$> d)
         submitYesVote_ (DRepVoter dRep) p116
-        submitYesVote_ (CommitteeVoter committeeMember1) p116
-        submitYesVote_ (CommitteeVoter committeeMember2) p116
+        submitYesVoteCCs_ committeeMembers' p116
         passNEpochs 3
         fmap (!! 3) getProposalsForest
           `shouldReturn` Node (SJust p116) []
@@ -937,12 +927,11 @@ constitutionSpec =
 
     describe "rejected for" $ do
       it "empty PrevGovId after the first constitution was enacted" $ do
-        (committeeMember1 :| [committeeMember2]) <- registerInitialCommittee
+        committeeMembers' <- registerInitialCommittee
         (dRep, _, _) <- setupSingleDRep 1_000_000
         (govActionId, _constitution) <- submitConstitution SNothing
         submitYesVote_ (DRepVoter dRep) govActionId
-        submitYesVote_ (CommitteeVoter committeeMember1) govActionId
-        submitYesVote_ (CommitteeVoter committeeMember2) govActionId
+        submitYesVoteCCs_ committeeMembers' govActionId
         passNEpochs 2
         constitution <- arbitrary
         let invalidNewConstitutionGovAction =